Fix CI compiler warnings and lint errors.
Drop unnecessary null assertions and a redundant Long conversion.
Disable UNCHECKED_CAST for App Functions KSP output.
Map MediaExtractor sample flags to MediaCodec buffer flags, and
read snackbar strings from LocalResources.

+++ app/src/main/java/com/androidcompress/app/encode/FfmpegKitGateway.kt
@@ -44,7 +44,7 @@ class FfmpegKitGateway : FfmpegGateway {
                 onStats(
                     EncodeStats(
                         timeMs = stats.time.toLong().coerceAtLeast(0L),
-                        sizeBytes = stats.size.toLong().coerceAtLeast(0L),
+                        sizeBytes = stats.size.coerceAtLeast(0L),
                         speed = stats.speed.toFloat(),
                         videoFrameNumber = stats.videoFrameNumber.coerceAtLeast(0),
                     ),
